# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Tracker API — Phase 3 + Wave 4 surface
  version: 0.6.0
  description: |
    Phase-3 redesign: the memory-card user-curation state machine has been
    removed. There is no `state` enum, no `needs_review` flag, no
    PATCH/commit/reject routes. Cards are live as soon as the agent creates
    them; quality is owned by the audit + consolidation passes which write
    `confidence`, `hidden_at`, `superseded_by_id`, and `audit_reason`.
    The list endpoint hides agent-soft-deleted / superseded rows by default.

    Wave 2 (this version) adds the agent quality endpoints:
      * PATCH /api/memory-cards/{id}/confidence — audit pass downgrades a card
      * POST  /api/memory-cards/{id}/hide       — audit pass soft-deletes a card
      * POST  /api/memory-cards/{loser}/supersede-into/{winner} — consolidation merge

    Wave 4 adds cross-meeting recall: Postgres FTS over speaker_segments
    (4.1) and memory_cards (4.2) under `/api/search/*`, plus a workspace-
    wide Hermes Q&A endpoint at `/api/qa/workspace` (4.3).
servers:
  - url: http://localhost:8000
paths:
  /api/meetings/{meeting_id}/import-job:
    parameters:
      - { name: meeting_id, in: path, required: true, schema: { type: string } }
    get:
      summary: Read import processing status and attempt count
      responses:
        '200':
          description: Current import job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportJob' }
        '404': { description: Missing, deleted or unsupported import }
  /api/meetings/{meeting_id}/import-job/cancel:
    parameters:
      - { name: meeting_id, in: path, required: true, schema: { type: string } }
    post:
      summary: Revoke local result publication; does not cancel remote computation
      responses:
        '200':
          description: Cancelled locally, including already-cancelled requests
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportJob' }
        '404': { description: Missing, deleted or unsupported import }
        '409': { description: Import already completed or failed }
  /api/meetings/{meeting_id}/import-job/retry:
    parameters:
      - { name: meeting_id, in: path, required: true, schema: { type: string } }
    post:
      summary: Explicitly requeue failed or cancelled imports
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                remote_stopped:
                  type: boolean
                  default: false
                  description: User verified that the previous remote job stopped; required if any prior attempt started.
      responses:
        '202':
          description: Persisted pending retry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportJob' }
        '404': { description: Missing, deleted or unsupported import }
        '409': { description: Not retryable or remote-stop confirmation missing }
        '422': { description: Invalid request body }
  /api/conversations/import:
    post:
      summary: Import a conversation (voice file, transcript file, or pasted text)
      description: |
        Single entrance for all imported conversations. The router (worktree D)
        inspects the request, persists a ConversationArtifact + Meeting,
        and dispatches to the voice-ingest pipeline (worktree B) or the
        transcript-ingest pipeline (worktree C).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: "Server enforces: exactly one of voice_file, transcript_file, or pasted_transcript MUST be provided. OpenAPI 3.0 cannot express this constraint at the schema level; the server returns 400 'no_input' if zero are provided and 400 'multiple_inputs' if more than one is provided."
              required: [workspace_id, title]
              properties:
                workspace_id:
                  type: string
                title:
                  type: string
                visibility:
                  type: string
                  enum: [private, workspace, shared]
                  default: private
                labels:
                  type: array
                  items: { type: string }
                voice_file:
                  type: string
                  format: binary
                  description: Audio file. Use this OR transcript_file OR pasted_transcript.
                transcript_file:
                  type: string
                  format: binary
                pasted_transcript:
                  type: string
      responses:
        '202':
          description: Accepted; processing begins asynchronously.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportAccepted'
        '400':
          description: "Invalid input. Codes: no_input (zero body parts present), multiple_inputs (more than one body part present), invalid_format (unsupported file type)."
...
components:
  schemas:
    ImportAccepted:
      type: object
      required: [artifact_id, meeting_id, processing_status]
      properties:
        artifact_id: { type: string }
        meeting_id:  { type: string }
        processing_status:
          type: string
          description: At 202 time the artifact is always 'received'. Subsequent state transitions are observed via GET /api/meetings/{id}.
          enum: [received]
    ImportJob:
      type: object
      required: [status, attempts, error]
      properties:
        status: { type: string }
        attempts: { type: integer, minimum: 0 }
        error: { type: string, nullable: true }
...